feat(infra): extract test modules from spine/forge prod files (#261 Moves 1b-d)
